+++ completion/utilities.py
@@ -10,9 +10,10 @@ from .exceptions import UnavailableCompletionData
 from .models import BlockCompletion
 
 
-def get_key_to_last_completed_course_block(user, course_key):
+def get_key_to_last_completed_block(user, context_key):
     """
-    Returns the last block a "user" completed in a course (stated as "course_key").
+    Returns the last block a "user" completed in a learning context such as a
+    specific course (stated as "context_key").
 
     raises UnavailableCompletionData when the user has not completed blocks in
     the course.
@@ -21,9 +22,9 @@ def get_key_to_last_completed_course_block(user, course_key):
     disabled.
     """
 
-    last_completed_block = BlockCompletion.get_latest_block_completed(user, course_key)
+    last_completed_block = BlockCompletion.get_latest_block_completed(user, context_key)
 
     if last_completed_block is not None:
         return last_completed_block.block_key
 
-    raise UnavailableCompletionData(course_key)
+    raise UnavailableCompletionData(context_key)